main_pytorch.py: change sample_init_ts_arr type from int to float.

+++ RectifiedFlow_Pytorch/main_pytorch.py
@@ -73,7 +73,7 @@ def parse_args_and_config():
     parser.add_argument("--sample_ckpt_path", type=str, default='./checkpoint/ckpt_gnobitab_RF_LSUN_Bedroom.pth')
     parser.add_argument("--sample_output_dir", type=str, default="./output5/generated")
     parser.add_argument("--sample_steps_arr", nargs='*', type=int, default=[10])
-    parser.add_argument("--sample_init_ts_arr", nargs='*', type=int, default=[940])
+    parser.add_argument("--sample_init_ts_arr", nargs='*', type=float, default=[940])
     parser.add_argument("--sample_isc_flag", type=str2bool, default=False, help="calculate IS for samples")
     parser.add_argument("--fid_input1", type=str, default="../ddim/exp/datasets/lsun/bedroom_train")
